Module tests.
